refactor(extension): relocate browser extension under apps/extension
Move the Plasmo extension into apps/extension and adopt @ybdownload/shared
for URLs, deep links, and YouTube URL validation.
