+++ apps/extension/.prettierignore
