+++ apps/extension/contents/components/FloatingButton.tsx
