+++ apps/extension/contents/lib/storage.ts
